Metadata: quiz roles visibility
